+++ packages/jaspr_content/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased minor
+
+- Fixed typo with `keepSuffixPattern` parameter in `FilesystemLoader` and `GithubLoader`.
+
 ## 0.1.1
 
 - `jaspr` upgraded to `0.19.0`